Create 30-7-2025.py
Today's problem required me to find out every possible subarray with some of it elements equal to a given number, here k, for a given array.

My code solves the problem but faces only one issue, i.e., for time optimization. What I found was that the sum() was O(^3), and I should've had something like O(n).

I didn't resolve this issue since i'm not aware of the concept as of now have it marked in my To-Learn-List!